analytics: Extract analytics/views/realm_activity.py.
